PRG: 37678, sort before offset/limit on entire memberlist
